app/requirements-to-freeze.txt
Flask
Flask-Babel
Flask-CORS
Flask-Login
Flask-Migrate
Flask-Principal
Flask-Script
Flask-SQLAlchemy
Flask-WTF
identity-toolkit-python-client
python-frontmatter
requests
uWSGI
